+++ packages/core/datetime-picker/src/internal/__tests__/integration/DateTimePicker.js
@@ -2,7 +2,8 @@
 import { BrowserTestCase } from '@atlaskit/webdriver-runner/runner';
 import { getExampleUrl } from '@atlaskit/webdriver-runner/utils/example';
 import Page from '@atlaskit/webdriver-runner/wd-wrapper';
-import * as assert from 'assert';
+// TODO: AK-5546: There is an issue with .log
+// import * as assert from 'assert';
 
 const urlDateTimePicker = getExampleUrl('core', 'datetime-picker', 'basic');
 /* Css used for the test */
@@ -28,15 +29,16 @@ BrowserTestCase(
     const previousDate = await dateTimePickerTest.getText(dateValue);
     await dateTimePickerTest.keys(['Backspace']);
     expect(await dateTimePickerTest.getText(dateValue)).not.toBe(previousDate);
-    if (dateTimePickerTest.log('browser').value) {
-      dateTimePickerTest.log('browser').value.forEach(val => {
-        assert.notEqual(
-          val.level,
-          'SEVERE',
-          `Console errors :${val.message} when the input is cleared`,
-        );
-      });
-    }
+    // TODO: AK-5546: There is an issue with .log
+    // if (dateTimePickerTest.log('browser').value) {
+    //   dateTimePickerTest.log('browser').value.forEach(val => {
+    //     assert.notEqual(
+    //       val.level,
+    //       'SEVERE',
+    //       `Console errors :${val.message} when the input is cleared`,
+    //     );
+    //   });
+    // }
   },
 );
 
@@ -55,15 +57,16 @@ BrowserTestCase(
     await dateTimePickerTest.keys(['ArrowLeft']);
     await dateTimePickerTest.keys(['Enter']);
     expect(await dateTimePickerTest.getText(dateValue)).not.toBe(previousDate);
-    if (dateTimePickerTest.log('browser').value) {
-      dateTimePickerTest.log('browser').value.forEach(val => {
-        assert.notEqual(
-          val.level,
-          'SEVERE',
-          `Console errors :${val.message} when the date is updated`,
-        );
-      });
-    }
+    // TODO: AK-5546: There is an issue with .log
+    // if (dateTimePickerTest.log('browser').value) {
+    //   dateTimePickerTest.log('browser').value.forEach(val => {
+    //     assert.notEqual(
+    //       val.level,
+    //       'SEVERE',
+    //       `Console errors :${val.message} when the date is updated`,
+    //     );
+    //   });
+    // }
   },
 );
 
@@ -82,14 +85,15 @@ BrowserTestCase(
     timePicker.keys(['Enter']);
     expect(await timePicker.getText(timeValue)).not.toBe(previousTime);
     expect(await timePicker.getText(timeValue)).toBe('12:45pm');
-    if (timePicker.log('browser').value) {
-      timePicker.log('browser').value.forEach(val => {
-        assert.notEqual(
-          val.level,
-          'SEVERE',
-          `Console errors :${val.message} when the time is updated`,
-        );
-      });
-    }
+    // TODO: AK-5546: There is an issue with .log
+    // if (timePicker.log('browser').value) {
+    //   timePicker.log('browser').value.forEach(val => {
+    //     assert.notEqual(
+    //       val.level,
+    //       'SEVERE',
+    //       `Console errors :${val.message} when the time is updated`,
+    //     );
+    //   });
+    // }
   },
 );