+++ packages/editor/adf-utils/validate.d.ts
@@ -1,1 +1,0 @@
-export { validate } from './dist/es5/validator';