+++ packages/editor/renderer/src/email/nodes/heading.ts
@@ -1,21 +1,21 @@
-import { akTypographyMixins } from '@atlaskit/util-shared-styles';
+import { typography } from '@atlaskit/theme';
 import { NodeSerializerOpts } from '../interfaces';
 import { createTag } from '../util';
 
 const getTypoMixin = (tagName: string) => {
   switch (tagName) {
     case 'h1':
-      return akTypographyMixins.h800;
+      return typography.h800();
     case 'h2':
-      return akTypographyMixins.h700;
+      return typography.h700();
     case 'h3':
-      return akTypographyMixins.h600;
+      return typography.h600();
     case 'h4':
-      return akTypographyMixins.h500;
+      return typography.h500();
     case 'h5':
-      return akTypographyMixins.h400;
+      return typography.h400();
     case 'h6':
-      return akTypographyMixins.h300;
+      return typography.h300();
     default:
       throw new Error(`Unknown tagName: ${tagName}`);
   }